requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
supabase==2.10.0
httpx>=0.26,<0.28
pydantic==2.10.3
pydantic-settings==2.7.0
python-multipart==0.0.20
pdfplumber==0.11.4
python-docx==1.1.2
python-jose[cryptography]==3.3.0
